requirements.txt
sqlmodel
textual
passlib[bcrypt]
email-validator
fpdf2
